Changed some functions names and logged some errors
git-svn-id: svn://svn.savannah.gnu.org/myserver/trunk@942 c4c922ea-a70d-0410-937d-ec1daf91936c

+++ myserverweb/include/processes.h
@@ -50,7 +50,7 @@ struct START_PROC_INFO
 };
 #endif
 
-u_long execHiddenProcess(START_PROC_INFO* spi,u_long timeout=0xFFFFFFFF);
-u_long execConcurrentProcess(START_PROC_INFO* spi);
+int execHiddenProcess(START_PROC_INFO* spi,u_long timeout=0xFFFFFFFF);
+int execConcurrentProcess(START_PROC_INFO* spi);
 int terminateProcess(u_long);
 #endif
